Rename folder name to align with convention
