+++ examples/camera-app/linux/include/clusters/webrtc-provider/webrtc-provider-manager.h
