+++ examples/camera-app/linux/src/clusters/webrtc-provider/webrtc-provider-manager.cpp
