+++ examples/camera-controller/webrtc-manager/WebRTCProviderClient.cpp
