+++ examples/camera-controller/webrtc-manager/WebRTCProviderClient.h
