+++ examples/camera-controller/webrtc-manager/WebRTCRequestorDelegate.cpp
