+++ examples/camera-controller/webrtc-manager/WebRTCRequestorDelegate.h
